Split a table in two and remove repeated values.

This is a subset of the tasks done by nest.

unjoin(data, ...)

# S3 method for unjoin
unjoin(data, ..., key_col = ".idx0")

# S3 method for data.frame
unjoin(data, ..., key_col = ".idx0")

Arguments

data

A data frame.

...

Specification of columns to nest. Use bare variable names. Select all variables between x and z with x:z, exclude y with -y. For more options, see the select documentation.

key_col

The name of the new column to key the two output data frames.

See also

`dplyr::inner_join` for the inverse operation.

`tidyr::nest` for the complementary operation resulting in one data frame

Examples

library(dplyr)
data("Seatbelts", package= "datasets")
x <- unjoin(as.data.frame(Seatbelts), front, rear, kms)
all.equal(inner_join(x$.idx0, x$data) %>% select(-.idx0), as.data.frame(Seatbelts))

#> Joining, by = ".idx0"

#> [1] TRUE

iris %>% unjoin(-Species)

#> $.idx0
#> # A tibble: 149 × 5
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width .idx0
#>           <dbl>       <dbl>        <dbl>       <dbl> <int>
#> 1           5.1         3.5          1.4         0.2    36
#> 2           4.9         3.0          1.4         0.2    19
#> 3           4.7         3.2          1.3         0.2    10
#> 4           4.6         3.1          1.5         0.2     6
#> 5           5.0         3.6          1.4         0.2    32
#> 6           5.4         3.9          1.7         0.4    52
#> 7           4.6         3.4          1.4         0.3     8
#> 8           5.0         3.4          1.5         0.2    28
#> 9           4.4         2.9          1.4         0.2     2
#> 10          4.9         3.1          1.5         0.1    20
#> # ... with 139 more rows
#> 
#> $data
#> # A tibble: 150 × 2
#>    Species .idx0
#>     <fctr> <int>
#> 1   setosa    36
#> 2   setosa    19
#> 3   setosa    10
#> 4   setosa     6
#> 5   setosa    32
#> 6   setosa    52
#> 7   setosa     8
#> 8   setosa    28
#> 9   setosa     2
#> 10  setosa    20
#> # ... with 140 more rows
#> 
#> attr(,"class")
#> [1] "unjoin"

chickwts %>% unjoin(weight)

#> $.idx0
#> # A tibble: 66 × 2
#>    weight .idx0
#>     <dbl> <int>
#> 1     179    14
#> 2     160    10
#> 3     136     3
#> 4     227    25
#> 5     217    22
#> 6     168    11
#> 7     108     1
#> 8     124     2
#> 9     143     6
#> 10    140     4
#> # ... with 56 more rows
#> 
#> $data
#> # A tibble: 71 × 2
#>         feed .idx0
#>       <fctr> <int>
#> 1  horsebean    14
#> 2  horsebean    10
#> 3  horsebean     3
#> 4  horsebean    25
#> 5  horsebean    22
#> 6  horsebean    11
#> 7  horsebean     1
#> 8  horsebean     2
#> 9  horsebean     6
#> 10 horsebean     4
#> # ... with 61 more rows
#> 
#> attr(,"class")
#> [1] "unjoin"

if (require("gapminder")) {
  gapminder %>%
    group_by(country, continent) %>%
    unjoin()

  gapminder %>%
    unjoin(-country, -continent)
  unjoin(gapminder)
}

#> $.idx0
#> # A tibble: 1 × 1
#>   .idx0
#>   <int>
#> 1     1
#> 
#> $data
#> # A tibble: 1,704 × 7
#>        country continent  year lifeExp      pop gdpPercap .idx0
#>         <fctr>    <fctr> <int>   <dbl>    <int>     <dbl> <int>
#> 1  Afghanistan      Asia  1952  28.801  8425333  779.4453     1
#> 2  Afghanistan      Asia  1957  30.332  9240934  820.8530     1
#> 3  Afghanistan      Asia  1962  31.997 10267083  853.1007     1
#> 4  Afghanistan      Asia  1967  34.020 11537966  836.1971     1
#> 5  Afghanistan      Asia  1972  36.088 13079460  739.9811     1
#> 6  Afghanistan      Asia  1977  38.438 14880372  786.1134     1
#> 7  Afghanistan      Asia  1982  39.854 12881816  978.0114     1
#> 8  Afghanistan      Asia  1987  40.822 13867957  852.3959     1
#> 9  Afghanistan      Asia  1992  41.674 16317921  649.3414     1
#> 10 Afghanistan      Asia  1997  41.763 22227415  635.3414     1
#> # ... with 1,694 more rows
#> 
#> attr(,"class")
#> [1] "unjoin"

unjoin(iris, Petal.Width)

#> $.idx0
#> # A tibble: 22 × 2
#>    Petal.Width .idx0
#>          <dbl> <int>
#> 1          0.2     2
#> 2          0.4     4
#> 3          0.3     3
#> 4          0.1     1
#> 5          0.5     5
#> 6          0.6     6
#> 7          1.4    11
#> 8          1.5    12
#> 9          1.3    10
#> 10         1.6    13
#> # ... with 12 more rows
#> 
#> $data
#> # A tibble: 150 × 5
#>    Sepal.Length Sepal.Width Petal.Length Species .idx0
#>           <dbl>       <dbl>        <dbl>  <fctr> <int>
#> 1           5.1         3.5          1.4  setosa     2
#> 2           4.9         3.0          1.4  setosa     2
#> 3           4.7         3.2          1.3  setosa     2
#> 4           4.6         3.1          1.5  setosa     2
#> 5           5.0         3.6          1.4  setosa     2
#> 6           5.4         3.9          1.7  setosa     4
#> 7           4.6         3.4          1.4  setosa     3
#> 8           5.0         3.4          1.5  setosa     2
#> 9           4.4         2.9          1.4  setosa     2
#> 10          4.9         3.1          1.5  setosa     1
#> # ... with 140 more rows
#> 
#> attr(,"class")
#> [1] "unjoin"